feat(model): remove JSON omitempty tags from message ID fields
Removed `omitempty` JSON tags from `ID` and `SessionID` fields in the `Message` struct to ensure these fields are always included in JSON serialization, even when they have zero values. This change guarantees consistent message representation in the API responses.

+++ internal/model/message.go
@@ -19,8 +19,8 @@ const (
 
 // Message represents a WhatsApp message, closely following whatsmeow structure
 type Message struct {
-	ID        int64 `json:"id"`
-	SessionID int   `json:"sessionId"`
+	ID        int `json:"id"`
+	SessionID int `json:"sessionId"`
 	
 	// Core identifiers (from whatsmeow MessageInfo)
 	MessageID string    `json:"messageId"` // Info.ID
